lib/liquid_crypto: update checks on aad length
Update checks on aad length

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/dataplane/dpu-offload/+/157502
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>

+++ lib/liquid_crypto/dao_liquid_crypto.h
@@ -86,7 +86,7 @@ struct dao_lc_sym_op {
 	/** AAD. Ignored for non-AEAD operations. */
 	uint8_t *aad;
 	/** AAD length. Ignored for non-AEAD operations. */
-	uint8_t aad_len;
+	uint16_t aad_len;
 	/** Digest. Ignored for non auth use cases. */
 	uint8_t *digest;
 	/** Operation. Whether the operation is Encrypt or Decrypt */
